Undo a few copyright / attribution mistakes
Some wrong copyright holder and attribution information was added
accidentally; this commit fixes that.  Specifically, this reverts part
of commit 6cb1ad7f (LICENSE.md), commit 393a31490 and part of earlier
commit e1deda29b (etl/leie/README.mdwn), and part of commit 63b6c8198a
(psm-app/userhelp/source/conf.py, though the file was located at
docs/userhelp/source/conf.py at the time of the original commit).

Note that correct copyright holder information will need to be
ascertained and added later.  In psm-app/userhelp/source/conf.py, the
"(TBD)" added here is a temporary workaround, because that copyright
metadata may be displayed as part of the documentation but we do not
yet have all the information needed to fill it in.