Dockerize leie module
Docker makes it easy to spin up a working copy of the leie module.  This
sets up the Dockerfile as well as a more general docker-compose file
(since we will eventually have more than one container associated with
this repository).

Since docker is a container, the flask server also needed to be modified
to listen to non-local traffic.